Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    double a;
+    float a;
     scanf("%f",&a);
-    printf("You entered: %.4f",a);
+    printf("You entered: %f",a);
     return 0;
 }
